apiSQUAD/requirements.txt
Flask
Flask-PyMongo
requests
flask-restful
flask-apispec==0.11.0
invoke